OffStage
Introduces a new Widget that lays its child out, but does not render it.
Also, moves RenderOverflowBox to its own file since it has nothing to do
with proxy box really.
